core/standards/dnd/synthetic/008-1.html

<!doctype html>
<html>
	<head>
		<title>Data persistence over multiple pages</title>
		<script type="text/javascript" src="/resources/testharness.js"></script>
		<script type="text/javascript" src="/resources/testharnessreport.js"></script>
	</head>
	<body>
		<div id="log"></div>
		<script type="text/javascript">
var div = document.createElement('div');
div.ondrop = function (e) {
	e.preventDefault();
	var dataTransfer = e.dataTransfer;
	test(function () {
		assert_equals( dataTransfer.getData('text'), '' );
	},'data store must not persist over multiple page loads');
};
var evt = document.createEvent('DragEvent');
evt.initDragEvent('drop', true, true, window, 1, 0, 0, 0, 0, false, false, false, false, 1, document.body, {});
div.dispatchEvent(evt);
		</script>
		<noscript><p>Enable JavaScript and reload</p></noscript>
	</body>
</html>
